finish

+++ Source/Engine/Common/Scene/Camera.cpp
@@ -23,7 +23,7 @@ void Camera::InitCamera(glm::vec3 Position, glm::vec3 Target, glm::vec3 UpDir)
                                    viewPoint + position,yAxis);
 
     this->vpMat.proj = glm::perspective(cameraData.fov,
-                                        VulkanContext::GetContext().windowExtent.width/(float)VulkanContext::GetContext().windowExtent.height,
+                                        (float)VulkanContext::GetContext().windowExtent.width/(float)VulkanContext::GetContext().windowExtent.height,
                                         cameraData.nearPlane, cameraData.farPlane);
     this->vpMat.proj[1][1] *=-1;
